extract interface for client
